# src/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Result": {
				"properties": {
					"status": {
						"type": "string"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"status",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "mock",
		"version": "0.0.0",
		"description": "TypeScript + Prisma + MySQL + Express API Server",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/status": {
			"get": {
				"operationId": "Index",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Result"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/mock/requests": {
			"put": {
				"operationId": "ResetMockRequestsApi",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"hash": {
										"type": "number",
										"enum": [
											1
										],
										"nullable": false
									}
								},
								"required": [
									"hash"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/{serverName}/mocks": {
			"get": {
				"operationId": "GetMocskApi",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "serverName",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/proxy": {
			"post": {
				"operationId": "CreateProxyApi",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"url": {
										"type": "string"
									},
									"name": {
										"type": "string"
									}
								},
								"required": [
									"url",
									"name"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/{serverName}/config": {
			"put": {
				"operationId": "RecordingApi",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "serverName",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"proxy": {
										"properties": {
											"omitHeaders": {
												"items": {
													"type": "string"
												},
												"type": "array"
											},
											"enabled": {
												"type": "boolean"
											}
										},
										"required": [
											"omitHeaders",
											"enabled"
										],
										"type": "object"
									},
									"serverUrl": {
										"type": "string"
									}
								},
								"required": [
									"proxy",
									"serverUrl"
								],
								"type": "object"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetConfigApi",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "serverName",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}